separated stacks
