+++ examples/decoupled-stacks/authservice-api-schema.json
@@ -1,0 +1,37 @@
+{
+    "openapi": "3.0.3",
+    "info": {
+      "title": "central-api-managed-api-gw",
+      "version": "1.0.0",
+      "description": "Acme Central API GW",
+      "x-reconciler": {
+        "pathPrefix": "authservice"
+      }
+    },
+    "paths": {},
+    "components": {
+
+      "securitySchemes": {
+        "UserTokenAuth": {
+          "type": "apiKey",
+          "name": "authorization",
+          "in": "header",
+          "x-amazon-apigateway-authtype": "custom",
+          "x-amazon-apigateway-authorizer": {
+            "type": "request",
+            "identitySource": "method.request.header.authorization",
+            "identityValidationExpression": "Bearer .*",
+            "authorizerUri": "authserviceOperationUri",
+            "authorizerResultTtlInSeconds": 0
+          }
+        }
+      }
+    },
+    "x-amazon-apigateway-request-validators": {
+      "basic": {
+        "validateRequestBody": true,
+        "validateRequestParameters": true
+      }
+    },
+    "x-amazon-apigateway-request-validator": "basic"
+  }
