+++ examples/decoupled-stacks/config.ts
@@ -1,0 +1,3 @@
+export const reconcilerMergedBucketArn="arn:aws:s3:::standalonereoncilerstack-reconcilermergedopenapid-1dl64grbno42y";
+export const reconcilerAPISchemaBucketArn= "arn:aws:s3:::standalonereoncilerstack-reconcileropenapidefinit-1drpnhtht049n";
+export const centralAPIGW="arn:aws:execute-api:ap-south-1:912469213006:1shn4k37a6/*/*/"
