+++ examples/new-rs-stack/config.ts
@@ -1,0 +1,6 @@
+export const reconcilerMergedBucketArn="arn:aws:s3:::newrsreoncilerstack-reconcilermergedopenapidefini-luupn91k5zze";
+export const reconcilerAPISchemaBucketArn= "arn:aws:s3:::newrsreoncilerstack-reconcileropenapidefinitionbu-pk18na1kpkrs";
+export const centralAPIGW="arn:aws:execute-api:ap-southeast-1:912469213006:66fvzlp5oc/*/*/"
+
+export const centralAPIGWId ="66fvzlp5oc";
+export const centralAPIGWrootResourceId= "rebpklsyce";
