+++ examples/new-rs-stack/readme.md
@@ -1,0 +1,5 @@
+export DEPLOY_ACCOUNT_ID="xxxxxxxxxx"  
+export DEPLOY_REGION=ap-southeast-1     
+
+cdk deploy --app 'npx ts-node reconciler.ts'
+
